Added bullet, HPsegment, and Main Menu
Main menu has a sample tank, so photoshop file is provided with layers to add the final tank when needed.

HP segment is only one part of the health bar. Black corner is added to make it easier to line up the segments 

Not much to say about bullet other then slightly asymmetrical, doubt it'll be a problem.